// apps/web/types/sanity.ts
export type Hero = {
  _type: "hero";
  slides: Array<{
    title?: string;
    subTitle?: string;
    buttons?: Array<
      {
        _key: string;
      } & Button
    >;
    mediaType?: "image" | "video";
    image?: {
      asset?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
      };
      media?: unknown;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      alt?: string;
      _type: "image";
    };
    video?: MuxVideo;
    _type: "heroSlide";
    _key: string;
  }>;
};

export type Button = {
  _type: "button";
  variant?:
    | "default"
    | "secondary"
    | "outline"
    | "link"
    | "ghost"
    | "destructive"
    | "fullGhost"
    | "CTA";
  text?: string;
  linkType: "internal" | "external";
  internalLink?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "page";
  };
  externalLink?: string;
  openInNewTab?: boolean;
};
// ...
export type MuxVideo = {
  _type: "mux.video";
  asset?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "mux.videoAsset";
  };
};
// ...
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;